Swap submissions list out for new component in dashboard
