+++ app/src/main/java/com/example/qstreak/ui/MainActivity.kt
@@ -34,12 +34,4 @@ class MainActivity : FragmentActivity(R.layout.activity_main) {
             .replace(R.id.fragment_container_view, fragment)
             .commit()
     }
-
-    fun navigateToTempDailyLog() {
-        val fragment = TempDailyLogFragment()
-        supportFragmentManager.beginTransaction()
-            .addToBackStack(TempDailyLogFragment.TAG)
-            .replace(R.id.fragment_container_view, fragment)
-            .commit()
-    }
 }
